- moved Add User modal and logic to AddUser component, clearing Admin Panel - removed RequestList
