+++ los-urlopos/src/components/AdminPanel/AdminPanel.tsx
@@ -1,26 +1,16 @@
 import styles from "./AdminPanel.module.css";
-import AddUser from "./AddUser/AddUser";
 import UsersList from "./UsersList/UsersList";
-import RequestsList from "./RequestsList/RequestsList";
 import DepartmentsList from "./DepartmentsList/DepartmentsList";
-import Modal from "react-modal";
 import BankHolidaysForm from "./BankHolidaysForm/BankHolidaysForm";
 import React, { FC, useState } from "react";
 
 const AdminPanel: FC = () => {
   const [selectedSection, setSelectedSection] = useState<string>("users");
-  const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-
-  const openModal = () => setIsModalOpen(true);
-  const closeModal = () => setIsModalOpen(false);
 
   const renderSection = () => {
     switch (selectedSection) {
       case "users":
-        return <UsersList openAddUserModal={openModal} />;
-
-      case "requests":
-        return <RequestsList />;
+        return <UsersList />;
       case "departments":
         return <DepartmentsList />;
       case "bankHolidays":
@@ -43,16 +33,9 @@ const AdminPanel: FC = () => {
         </button>
       </div>
       {renderSection()}
-      <Modal
-        isOpen={isModalOpen}
-        onRequestClose={closeModal}
-        contentLabel="Add User"
-      >
-        <button onClick={closeModal}>Close</button>
-        <AddUser onUserAdded={closeModal} />
-      </Modal>
     </div>
   );
 };
 
 export default AdminPanel;
+d;
