--- los-urlopos/src/components/AdminPanel/RequestsList/RequestsList.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const RequestsList = () => {
-  return <div>RequestsList</div>;
-};
-
-export default RequestsList;
