+++ los-urlopos/src/components/AdminPanel/UsersList/UsersList.tsx
@@ -3,18 +3,24 @@ import { User } from "../../../types-obj/types-obj";
 import styles from "./UsersList.module.css";
 import ConfirmAction from "../ConfirmAction";
 import { deleteUser, subscribeToUsers } from "../../../services/UserService";
+import AddUser from "../AddUser/AddUser";
+import Modal from "react-modal";
 
 interface UsersListProps {
   openAddUserModal: () => void;
 }
 
-const UsersList: FC<UsersListProps> = ({ openAddUserModal }) => {
+const UsersList: FC<UsersListProps> = ({}) => {
   const [users, setUsers] = useState<User[]>([]);
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<string | null>(null);
   const [expandedUserId, setExpandedUserId] = useState<string | null>(null);
   const [confirmDialogOpen, setConfirmDialogOpen] = useState<boolean>(false);
   const [userIdToDelete, setUserIdToDelete] = useState<string | null>(null);
+  const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
+
+  const openModal = () => setIsModalOpen(true);
+  const closeModal = () => setIsModalOpen(false);
 
   useEffect(() => {
     const unsub = subscribeToUsers(
@@ -69,7 +75,15 @@ const UsersList: FC<UsersListProps> = ({ openAddUserModal }) => {
   return (
     <div>
       <h1>Users</h1>
-      <button onClick={openAddUserModal}>Add User</button>
+      <button onClick={openModal}>Add User</button>
+      <Modal
+        isOpen={isModalOpen}
+        onRequestClose={closeModal}
+        contentLabel="Add User"
+      >
+        <button onClick={closeModal}>Close</button>
+        <AddUser onUserAdded={closeModal} />
+      </Modal>
       <ul>
         {users.map((user) => (
           <li key={user.id}>
